Fix import issues introduced by the refactor
