+++ src/kinesis-firehose/constructs/lib/buffering-configuration.ts
@@ -1,7 +1,7 @@
 import { Duration } from 'aws-cdk-lib';
 import { CfnDeliveryStream } from 'aws-cdk-lib/aws-kinesisfirehose';
 import { IConstruct } from 'constructs';
-import { definedFieldsOrUndefined } from '../../../../utils/formatting';
+import { definedFieldsOrUndefined } from '../../../utils/formatting';
 
 
 export interface BufferingConfigurationOptions {
