+++ src/kinesis-firehose/constructs/lib/destinations/s3-destination.ts
@@ -4,7 +4,7 @@ import { CfnDeliveryStream } from 'aws-cdk-lib/aws-kinesisfirehose';
 import { IKey } from 'aws-cdk-lib/aws-kms';
 import { IBucket } from 'aws-cdk-lib/aws-s3';
 import { IConstruct } from 'constructs';
-import { definedFieldsOrUndefined } from '../../../../../utils/formatting';
+import { definedFieldsOrUndefined } from '../../../../utils/formatting';
 import { BackupConfigurationResult, IDeliveryStreamBackupDestination } from '../backup-configuration';
 import { BufferingConfiguration } from '../buffering-configuration';
 import { CloudWatchLoggingConfiguration } from '../cloudwatch-logging-configuration';
